Major overhaul
* Split the JS code up, I'm now using webpack to package it
* Removed the loadTemplate mechanism and replaced it with vue-loader
* Updated the releng scripts to support node
* Updated configure.py accordingly
* Removed dynamic file lists from configure.py because they caused too
  many problems with file renames.
* Removed the obsolete converter

+++ releng/windows/auto-build.bat
@@ -1,16 +1,16 @@
 @echo off
 
 cd %~dp0..\..
-set "PATH=%PATH%;%CD%\releng\windows\support"
+set "PATH=%CD%\releng\windows\support;%PATH%"
 
-rmdir /S /Q releng\windows\dist
+if exist releng\windows\dist rmdir /S /Q releng\windows\dist
 mkdir releng\windows\dist
 
-echo ==> Configuring...
+echo ==^> Configuring...
 py -3 configure.py
 if errorlevel 1 exit /b 1
 
-echo ==> Building...
+echo ==^> Building...
 
 :build
 ninja installer
